pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "financial-compliance-rag"
version = "0.1.0"
description = "Financial Compliance and Risk RAG System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "langchain>=0.3",
    "langgraph>=0.2",
    "pinecone>=3.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "ragas>=0.2",
    "openai>=1.55",
    "tiktoken>=0.8",
    "tabula-py>=2.10",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
]

[tool.hatch.build.targets.wheel]
packages = ["ingestion", "rag_agent", "api", "eval"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["eval"]
